fix(referral): wire bonus service, fix schema, tests green
- Wire ReferralBonusService into PaymentRouter via build_payment_router
- Fix ROUND(balance + bonus) type cast for PostgreSQL real columns
- Align schema: reward_value DECIMAL(10,2), UNIQUE(referred_tg_id)
- Add migration 015 for existing production databases
- Fix pre-existing bot tests (circuit breaker, admin getters, registration middleware)
- Add backend unit/integration tests for referral bonus flow

+++ services/core/referral/bonus_service.py
@@ -233,7 +233,7 @@ class ReferralBonusService:
                 await conn.execute(
                     """
                     UPDATE users
-                    SET balance = ROUND(balance + $1, 2)
+                    SET balance = ROUND((balance + $1)::numeric, 2)
                     WHERE tg_id = $2
                     """,
                     reward_value, referrer_tg_id,
